+++ CHANGELOG.md
@@ -12,6 +12,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 
 
+## [?.?.?] - ????-??-??
+
+### Changed
+- Migrated to vitest.
+
+
 ## [0.17.1] - 2026-02-05
 
 ### Changed